backend/requirements.txt
asyncio
fastapi
httpx
pydantic
sqlalchemy
uvicorn
sqlmodel
alembic
pyyaml